Do component integer size optimalization

+++ src/ECS/Components/animationComponents.hpp
@@ -1,11 +1,6 @@
 #pragma once
 
 #include "Resources/animationStatesResources.hpp"
-#include "Utilities/rect.hpp"
-#include <SFML/System/Time.hpp>
-#include <map>
-#include <string>
-#include <vector>
 
 namespace ph::component {
 
@@ -15,7 +10,7 @@ struct AnimationData // @no-debugger
 	AnimationStatesData* states;
 	float delay = 0.1f;
 	float elapsedTime = 0.f;
-	u32 currentFrameIndex = 0;
+	u16 currentFrameIndex = 0;
 	bool isPlaying = false;
 };
